Developed Non-Uniform velocity
Behavior weights are normalized to 1 and now each component of the velocity function is multiplied by the scalar speed value vice multiplying scalar speed value after normalizing velocity. This will ensure the magnitude of the velocity is always equal to or less than the speed scalar (maximum speed) of the buoy. Also updated broadcast data to send velocity and speed.