Developed Target Class
Developed a target class that moves and affects the scalar of the environment. Moved environment instance generation into Swarm class so that a single environment is generated for use everywhere. Also moved measure method from buoys to swarm to help with organizing operating loop. Target class can be turned on or off as need for static environment testing or dynamic target tracking testing.