Stale_memory
Stale_memory developed but seems to break seeker performance. They converge on a location and do not move after that. Memory and best known parameters are updating as prescribed but does not work nearly as well as a fixed duration memory reset. Halting development on this feature until I can figure out why. Not folding progress into swarm master. However, target now has a circle vector to allow it to circle about the environment.